+++ src/mem/cache/tags/compressed_tags.test.cc
@@ -69,7 +69,7 @@ class SuperBlkTestFixture : public ::testing::Test
     verifyInvariants(const SuperBlk &sb)
     {
         uint8_t count_valid = 0;
-        uint8_t min_cf = sb.blks.size();
+        uint8_t min_cf = static_cast<uint8_t>(sb.blks.size());
         std::size_t total_bits = 0;
         for (const auto &blk : sb.blks) {
             if (blk->isValid()) {
@@ -189,7 +189,7 @@ TEST_F(SuperBlkTestFixture, SubBlockMigration)
 
     // Verify subBlksB[1] is valid and retained its 128-bit size
     ASSERT_TRUE(subBlksB[1].isValid());
-    ASSERT_EQ(subBlksB[1].getSizeBits(), 128);
+    ASSERT_EQ(subBlksB[1].getSizeBits(), 128U);
     ASSERT_EQ(superBlkB.getNumValid(), 1);
     ASSERT_EQ(superBlkB.getCompressionFactor(), 4);
